fixed style issues
