+++ src/Layout/Dashboard/Header.tsx
@@ -1,36 +1,13 @@
 import React, { useState } from "react";
-import { styled } from "@mui/material/styles";
+
 import { Box, Toolbar, Typography, IconButton, Badge } from "@mui/material";
-import MuiAppBar, { AppBarProps as MuiAppBarProps } from "@mui/material/AppBar";
 import MenuIcon from "@mui/icons-material/Menu";
 import AccountCircle from "@mui/icons-material/AccountCircle";
 import NotificationsIcon from "@mui/icons-material/Notifications";
 
 import { SearchBar } from "../../shared-components";
 import DrawerMenu from "./DrawerMenu";
-const drawerWidth = 240;
-
-interface AppBarProps extends MuiAppBarProps {
-  open?: boolean;
-}
-
-const AppBar = styled(MuiAppBar, {
-  shouldForwardProp: (prop) => prop !== "open",
-})<AppBarProps>(({ theme, open }) => ({
-  zIndex: theme.zIndex.drawer + 1,
-  transition: theme.transitions.create(["width", "margin"], {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  ...(open && {
-    marginLeft: drawerWidth,
-    // width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(["width", "margin"], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  }),
-}));
+import { AppBar } from "./style";
 
 const Header = () => {
   const [open, setOpen] = useState<boolean>(false);
@@ -53,10 +30,6 @@ const Header = () => {
             aria-label="open drawer"
             onClick={handleDrawerOpen}
             edge="start"
-            // sx={{
-            //   marginRight: 5,
-            //   // ...(open && { display: "none" }),
-            // }}
           >
             <MenuIcon />
           </IconButton>
